Added class EnemyDetectTargetLogic
Added class EnemyDetectTargetLogic
