+++ Assets/Scripts/Enemy/EnemyDetectTargetLogic.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 88ad53f1cb67e2c45a02716e1f527c15
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
