+++ Assets/Scripts/Enemy/EnemyStates/EnemyDetectTargetState.cs
@@ -1,8 +1,6 @@
-using UnityEngine;
-
 public class EnemyDetectTargetState : IState<BasicEnemy> {
 
-    private BasicEnemy _enemy;
+    protected EnemyDetectTargetLogic _enemyDetectLogic;
 
     public virtual void Enter(BasicEnemy owner) {
     }
@@ -11,73 +9,10 @@ public class EnemyDetectTargetState : IState<BasicEnemy> {
 
     }
 
-    public void CheckHasTargetAndChangeToIdleState(BasicEnemy enemy) {
-        if (!enemy.HasTarget) {
-            enemy.StateMachine.ChangeState(enemy.IdleState);
-            return;
-        }
-    }
-
-    public void CheckNeedFlipAndFlip(BasicEnemy enemy) {
-        if (enemy.IsNeedLookOnPlayer()) {
-            enemy.Flip();
-        }
-    }
-
-    public void CalculationDistanceToTarget(BasicEnemy enemy) {
-        if (enemy.FieldOfView.Target != null) {
-            enemy.distanceToTarget = Mathf.Abs(enemy.transform.position.x - enemy.FieldOfView.Target.transform.position.x);
-        }
-    }
-
-    public void CheckIsThereTargetInRangeOfAttackAndAttackOrRun(float distance, BasicEnemy enemy, EnemyAttackState attackState) {
-
-        if (enemy.IsThereTargetInRangeOfDistance(distance)) {
-            enemy.ResetRigidbodyVelocity();
-            enemy.Animator.Play(AnimationName.Idle);
-
-            HandleAttackDelay(enemy, attackState);
-        }
-        else {
-            enemy.Animator.Play(AnimationName.Run);
-        }
-    }
-
-    private void HandleAttackDelay(BasicEnemy enemy, EnemyAttackState attackState) {
-
-        enemy.Config.delayAttack -= Time.deltaTime;
-
-        if (enemy.Config.delayAttack <= 0) {
-            enemy.StateMachine.ChangeState(attackState);
-        }
-    }
-
-    public void ChangeToStateAttackAfterDelay(BasicEnemy enemy, IState<BasicEnemy> attackState, float delayAttack) {
-
-        enemy.ResetRigidbodyVelocity();
-        enemy.Animator.Play(AnimationName.Idle);
-
-        if (delayAttack <= 0) {
-            enemy.StateMachine.ChangeState(attackState);
-            return;
-        }
-    }
-
     public virtual void FixedUpdate() {
         
     }
 
-    public void MoveIfPlayRunAnimation(BasicEnemy enemy) {
-        if (enemy.Animator.GetCurrentAnimatorStateInfo(AnimatorLayers.BaseLayer).IsName(AnimationName.Run)) {
-            enemy.Move(enemy.GetDirectionX);
-        }
-    }
-
     public virtual void Exit() {
-    }
-
-    public void RefreshDelayForDifferentAttacks(BasicEnemy enemy) {
-        enemy.ConfigBuffer.delayAttack = enemy.Config.delayAttack;
-        enemy.ConfigBuffer.delayStrikeAttack = enemy.Config.delayStrikeAttack;
     }
 }
